clickhouseaware/requirements.txt
google-adk==0.5.0 
clickhouse_connect==0.8.17
jinja2==3.1.6
litellm==1.70.0
google-generativeai==0.3.2